requirements.txt
arrow==0.14.7; python_version < '3.0'
arrow==1.2.3; python_version >= '3.0' and python_version <= '3.7'
arrow==1.3.0; python_version >= '3.8'
bandit==1.6.2; python_version < '3.0'
bandit==1.7.1; python_version == '3.6'
bandit==1.7.5; python_version == '3.7'
bandit==1.7.10; python_version == '3.8'
bandit==1.8.6; python_version >= '3.9'
boto==2.49.0
boto3==1.17.112; python_version < '3.0'
boto3==1.23.10; python_version < '3.7' and python_version >= '3.0'
boto3==1.33.13; python_version == '3.7'
boto3==1.37.38; python_version == '3.8'
boto3==1.40.36; python_version >= '3.9'
Django==1.11.29; python_version < '3.0'
Django==3.2.25; python_version >= '3.0' and python_version <= '3.7'
Django==4.2.24; python_version >= '3.8' and python_version <= '3.9'
Django==5.2.6; python_version >= '3.10'
dropbox==12.0.2
google-api-python-client==2.182.0; python_version >= '3.7'
google-api-python-client==2.52.0; python_version >= '3.0' and python_version < '3.7'
google-api-python-client==1.12.11; python_version < '3.0'
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2; python_version >= '3.0'
google-auth-oauthlib==0.4.1; python_version < '3.0'
lockfile==0.12.2
paramiko==2.12.0; python_version < '3.0'
paramiko==3.5.1; python_version >= '3.0' and python_version < '3.10'
paramiko==4.0.0; python_version >= '3.10'
psycopg2-binary==2.8.6; python_version < '3.0'
psycopg2-binary==2.9.8; python_version >= '3.0' and python_version < '3.7'
psycopg2-binary==2.9.9; python_version == '3.7'
psycopg2-binary==2.9.10; python_version >= '3.8'
pylint==3.3.8; python_version >= '3.9'
pylint==3.2.7; python_version == '3.8'
pylint==2.17.7; python_version == '3.7'
pylint==2.13.9; python_version == '3.6'
pylint==1.9.5; python_version < '3.0'
PyNaCl==1.4.0; python_version < '3.0'
PyNaCl==1.5.0; python_version >= '3.0' and python_version < '3.8'
PyNaCl==1.6.0; python_version >= '3.8'
python-dotenv==0.18.0; python_version < '3.0'
python-dotenv==0.20.0; python_version >= '3.0' and python_version < '3.7'
python-dotenv==0.21.1; python_version == '3.7'
python-dotenv==1.0.1; python_version == '3.8'
python-dotenv==1.1.1; python_version >= '3.9'
pytz==2025.2
six==1.17.0
stone==3.3.1; python_version < '3.0'
